fix(molecular_helpers): cast sqrt result to f64 to resolve ambiguity

+++ crates/physics_engine/src/molecular_helpers.rs
@@ -118,7 +118,7 @@ impl PhysicsEngine {
         };
         
         // Lorentz-Berthelot mixing rules
-        let epsilon_mixed: f64 = (epsilon1 * epsilon2).sqrt();
+        let epsilon_mixed: f64 = (epsilon1 * epsilon2).sqrt() as f64;
         let sigma_mixed = (sigma1 + sigma2) / 2.0;
         
         (epsilon_mixed, sigma_mixed)
